Make configurable libvirt memory work again.
We expected to get a value in megabytes from the
configuration file.  Libvirt expects a value in kilobytes.
Therefore, we were taking the value from the configuration
and multiplying by 1024.  However, I forgot that
config_get_key() returns strings, not integers.  Multiplying
a string by a number gives you that many copies of the
string, which is not what we want.  Instead, cast to
an integer before multiplying.

+++ oz/Guest.py
@@ -163,8 +163,8 @@ class Guest(object):
                                                      1)
         # the memory in the configuration file is specified in megabytes, but
         # libvirt expects kilobytes, so multiply by 1024
-        self.install_memory = oz.ozutil.config_get_key(config, 'libvirt',
-                                                       'memory', 1024) * 1024
+        self.install_memory = int(oz.ozutil.config_get_key(config, 'libvirt',
+                                                           'memory', 1024) * 1024)
 
         # configuration from 'cache' section
         self.cache_original_media = oz.ozutil.config_get_boolean_key(config,
